Sliding Window Analysis for Time-Series Data

This function performs a sliding window analysis over a given data frame for the specified variable. It calculates the mean, sum, or count above a given threshold for the variable values in each window. The window size is defined by the user-specified window lengths.

Usage

windowpane(
  data,
  end_date,
  variable,
  summary_type,
  threshold = NULL,
  window_lengths
)

Arguments

data

A data frame that contains the data for analysis. It must include a YYYYMMDD date column.

end_date

A character string that specifies the end date for the analysis in the format 'YYYY-MM-DD'.

variable

The name of the variable in the data frame on which to perform the analysis.

summary_type

The type of summary statistic to compute for each window. Can be 'mean', 'sum', or 'above_threshold'.

threshold

An optional numerical value. If summary_type is 'above_threshold', the function counts how many variable values in each window are above this threshold.

window_lengths

A numerical vector that specifies the lengths of the windows for the analysis.

Value

A data frame in wide format where each column corresponds to a window of a certain length and start day, and contains the calculated summary statistic for that window. The column names are in the format 'variable_startday_endday'.

Examples

if (FALSE) {
data <- tibble(YYYYMMDD = seq(as.Date("2023-01-01"), as.Date("2023-12-31"), by = "day"),
               variable = rnorm(365))
windowpane(data, "2023-12-31", variable, "mean", NULL, c(7, 14, 21))
}