Test for Spatial Join Count Statistics

The function join_count calculates spatial statistics for a matrix. It identifies patterns of aggregation for values in a binary matrix based on join count statistics. The results determine whether the observed spatial arrangement is aggregated, random, or uniform regarding the presence of 1s ("D") and 0s ("H").

Usage

join_count(matrix_data)

Arguments

matrix_data

A binary matrix (with elements 0 and 1) representing the spatial distribution of two types of points. Rows and columns of the matrix indicate coordinates while the 0s and 1s represent different categories of observations.

Value

A list containing several statistical measures:

Observed_01_10: The number of observed 01 or 10 sequences.

Observed_11: The number of observed 11 sequences.

Expected_ER_01: The expected number of 01 or 10 sequences under random distribution.

Expected_ER_11: The expected number of 11 sequences under random distribution.

sR_01: Standard deviation for 01 or 10 sequences.

sR_11: Standard deviation for 11 sequences.

ZHD: The Z-score for 01 or 10 sequences.

ZDD: The Z-score for 11 sequences.

PatternHD: Descriptive result for 01 or 10 sequences - "Aggregated" or "Not Aggregated".

PatternDD: Descriptive result for 11 sequences - "Aggregated" or "Not Aggregated".

Details

The function first calculates the count of specific binary sequences in the matrix both horizontally and vertically. It then computes expected values and standard deviations based on the spatial arrangement dimensions and uses these for Z-score calculations to identify the pattern of the matrix.

Examples

matrix_data <- matrix(c(1,1,1,0,0,
                        1,1,1,0,0,
                        1,1,1,0,0,
                        1,1,1,0,0,
                        0,0,0,0,0), ncol = 5, byrow = TRUE)
join_count(matrix_data)
#> $Observed_01_10
#> [1] 7
#> 
#> $Observed_11
#> [1] 17
#> 
#> $Expected_ER_01
#> [1] 19.968
#> 
#> $Expected_ER_11
#> [1] 9.216
#> 
#> $sR_01
#> [1] 3.174123
#> 
#> $sR_11
#> [1] 4.231325
#> 
#> $ZHD
#> [1] -4.243062
#> 
#> $ZDD
#> [1] 1.957779
#> 
#> $PatternHD
#> [1] "Aggregated"
#> 
#> $PatternDD
#> [1] "Aggregated"
#>